feat: add function print to textarea from virtual keyboard, add style and animation to keys
