--- src/js/get_virtualkeyboard_value.js
@@ -1,8 +1,0 @@
-import {createKey} from "./create_key";
-
-export function getVirtualKeyboardValue() {
-    // let keys = document.getElementById(".keyboard__keys")
-   
-}
-
-getVirtualKeyboardValue();
