+++ style.css
@@ -6,8 +6,8 @@
     bottom: 0;
     padding: 10px 0;
     width: 100%;
-    background-color: #786f6e;
-    
+    background-color: #969190;
+    border-top: solid black 2px;    
 }
 
 .keyboard__keys {
@@ -26,15 +26,39 @@
     outline: none;
     cursor: pointer;
     font-size: large;
+    box-shadow: 6px 5px 5px 0px rgba(0,0,0,0.5);
 }
+
+.keyboard__key:hover {
+    background-color:  #c9c3c2;
+}
+
+.keyboard__key:active {
+    height: 42px;
+    width: 42px; 
+    margin: 6px;  
+    font-size: medium;    
+} 
 
 .keyboard__key_wide {
     width: 100px;
 }
 
+.keyboard__key_wide:active {
+    width: 96px;
+    margin: 6px;
+}
+
+
 .keyboard__key_space {
     width: 370px;   
 }
+
+.keyboard__key_space:active {
+    width: 366px;  
+    margin: 6px; 
+}
+
 
 .wrapper-textarea {
     position: fixed;
@@ -42,7 +66,7 @@
     top: 0;  
     width: 100%;
     height: 100vh;
-    background-color: cadetblue;
+    background-color: rgb(75, 69, 69);
     display: flex;
     justify-content: center;
     align-items: flex-start;
@@ -55,4 +79,7 @@
     height: 280px;
     margin: 15px 0px 0px;
     padding: 15px;
+    resize: none;
+    font-size: x-large;    
+    border: solid black 1.5px;
 }
